# swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "FIUS Pizza Calculator",
    "description": "Organize your pizza orders",
    "version": "1.0.0"
  },
  "basePath": "/",
  "paths": {
    "/js/filename": {
      "get": {
        "summary": "GET for every file in /js folder",
        "consumes": [
          "application/json"
        ],
        "tags": [
          "js/filename"
        ],
        "parameters": [],
        "responses": {
          "200": {
            "description": "file resource",
            "schema": {
              "type": "file"
            }
          }
        }
      }
    },
    "/node_modules/bootstrap/dist/css/filename": {
      "get": {
        "summary": "GET for every file in /bootstrap/dist/css folder",
        "consumes": [
          "application/json"
        ],
        "parameters": [],
        "tags": [
          "bootstrap",
          "css"
        ],
        "responses": {
          "200": {
            "description": "file resource",
            "schema": {
              "type": "file"
            }
          }
        }
      }
    },
    "/node_modules/bootstrap/dist/js/filename": {
      "get": {
        "summary": "GET for every file in /bootstrap/dist/js folder",
        "consumes": [
          "application/json"
        ],
        "parameters": [],
        "tags": [
          "bootstrap",
          "js"
        ],
        "responses": {
          "200": {
            "description": "file resource",
            "schema": {
              "type": "file"
            }
          }
        }
      }
    },
    "/teams": {
      "get": {
        "summary": "Show team page",
        "parameters": [],
        "tags": [
          "Teams"
        ],
        "responses": {
          "200": {
            "description": "file resource",
            "schema": {
              "type": "file"
            }
          }
        }
      }
    },
    "/pizzas/templates": {
      "get": {
        "summary": "Get all templates saved in file",
        "parameters": [],
        "tags": [
          "Pizza",
          "Template"
        ],
        "responses": {
          "200": {
            "description": "Array of available templates",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/template"
              }
            }
          },
          "404": {
            "description": "Empty body"
          }
        }
      },
      "post": {
        "summary": "Post a template",
        "parameters": [
          {
            "name": "template",
            "in": "body",
            "description": "Template",
            "required": true,
            "schema": {
              "$ref": "#/definitions/template"
            }
          }
        ],
        "tags": [
          "Pizzas",
          "Template"
        ],
        "responses": {
          "200": {
            "description": "Empty body if successfull"
          },
          "420": {
            "description": "Empty body if failed"
          }
        }
      }
    },
    "/pizzas/amount": {
      "get": {
        "summary": "Get amout of pizzas",
        "parameters": [
          {
            "name": "teamname",
            "in": "body",
            "description": "Teamname",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "tags": [
          "Pizza",
          "Amount"
        ],
        "responses": {
          "200": {
            "description": "Amount of pizza",
            "schema": {
              "$ref": "#/definitions/amount"
            }
          },
          "303": {
            "description": "Redirect to main page"
          }
        }
      }
    },
    "/pizzas/ingredients": {
      "get": {
        "summary": "Array of available ingredients",
        "parameters": [],
        "tags": [
          "Ingredients",
          "Pizzas"
        ],
        "responses": {
          "200": {
            "description": "Array of available ingredients",
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          }
        }
      }
    },
    "/pizzas/suggestions": {
      "get": {
        "summary": "Get all team suggestions",
        "consumes": [
          "application/json"
        ],
        "tags": [
          "Pizzas",
          "Suggestions"
        ],
        "parameters": [
          {
            "name": "teamname",
            "in": "query",
            "description": "Teamname",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Suggestions of a team",
            "schema": {
              "$ref": "#/definitions/suggestion"
            }
          },
          "303": {
            "description": "Redirect to main page"
          }
        }
      },
      "post": {
        "summary": "Add suggestion for a team",
        "parameters": [
          {
            "name": "suggestion",
            "in": "body",
            "description": "Suggestion",
            "required": true,
            "schema": {
              "$ref": "#/definitions/suggestionsParam"
            }
          }
        ],
        "tags": [
          "Pizzas"
        ],
        "responses": {
          "200": {
            "description": "Suggestion",
            "schema": {
              "$ref": "#/definitions/suggestion"
            }
          },
          "303": {
            "description": "Redirect to main page"
          }
        }
      }
    },
    "/teams/data/": {
      "get": {
        "summary": "Get team data",
        "parameters": [
          {
            "name": "teamname",
            "in": "query",
            "description": "Teamname",
            "required": true,
            "type": "string"
          }
        ],
        "tags": [
          "Teams"
        ],
        "responses": {
          "200": {
            "description": "Get team data",
            "schema": {
              "$ref": "#/definitions/teamData"
            }
          },
          "303": {
            "description": "Redirect to main page"
          }
        }
      }
    },
    "/teams/create": {
      "post": {
        "summary": "Set input field",
        "parameters": [
          {
            "name": "teamname",
            "in": "body",
            "description": "Teamname",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "tags": [
          "Teams"
        ],
        "responses": {
          "200": {
            "description": "Proposer file",
            "schema": {
              "type": "file"
            }
          },
          "420": {
            "description": "HTML Code",
            "schema": {
              "type": "string"
            }
          }
        }
      }
    },
    "/teams/teamsize": {
      "post": {
        "summary": "Set teamsize",
        "parameters": [
          {
            "name": "teamsizeParam",
            "in": "body",
            "description": "teamsizeParam",
            "required": true,
            "schema": {
              "$ref": "#/definitions/teamsizeParam"
            }
          }
        ],
        "tags": [
          "Pizzas",
          "Teams"
        ],
        "responses": {
          "302": {
            "description": "Redirect to subpage"
          },
          "303": {
            "description": "Redirect to main page"
          }
        }
      }
    },
    "/pizzas/suggestions/vote": {
      "post": {
        "summary": "Vote a suggestion",
        "parameters": [],
        "tags": [
          "Pizzas",
          "Suggestions",
          "Vote"
        ],
        "responses": {
          "200": {
            "description": "Empty body"
          },
          "303": {
            "description": "Redirect to main page"
          }
        }
      }
    }
  },
  "definitions": {
    "amount": {
      "type": "object",
      "properties": {
        "amount": {
          "type": "number"
        }
      }
    },
    "suggestion": {
      "type": "object",
      "properties": {
        "id": {
          "type": "number"
        },
        "ingredients": {
          "type": "array",
          "items": {
            "type": "string"
          }
        },
        "vote": {
          "type": "number"
        }
      }
    },
    "suggestionsParam": {
      "type": "object",
      "properties": {
        "teamname": {
          "type": "string"
        },
        "ingredients": {
          "type": "array",
          "items": {
            "type": "string"
          }
        }
      }
    },
    "teamsizeParam": {
      "type": "object",
      "properties": {
        "teamname": {
          "type": "string"
        },
        "count": {
          "type": "number"
        },
        "type": {
          "type": "string"
        }
      }
    },
    "teamData": {
      "type": "object",
      "properties": {
        "teamsize": {
          "$ref": "#/definitions/teamsize"
        },
        "count": {
          "type": "number"
        }
      }
    },
    "teamsize": {
      "type": "object",
      "properties": {
        "number": {
          "type": "number"
        },
        "type": {
          "type": "string"
        }
      }
    },
    "template": {
      "type": "object",
      "properties": {
        "name": {
          "type": "string"
        },
        "ingredients": {
          "type": "array",
          "items": {
            "type": "string"
          }
        }
      }
    }
  }
}